requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pymysql==1.1.0
cryptography==41.0.5
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
python-multipart==0.0.6